Move some utils to libft
