+++ libft/Makefile
@@ -43,7 +43,9 @@ SRCS = ft_atoi.c \
 	   ft_lstmap.c \
 	   ft_lstnew.c \
 	   ft_lstsize.c \
-	   ft_strncpy.c
+	   ft_strncpy.c \
+	   ft_strcmp.c \
+	   ft_strcpy.c
 
 OBJS = $(SRCS:.c=.o)
