+++ libft/ft_strcpy.c
@@ -1,0 +1,24 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_strcpy.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: vd-ambro <vd-ambro@student.42roma.it>      +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/11/27 11:44:33 by vd-ambro          #+#    #+#             */
+/*   Updated: 2023/11/27 11:45:09 by vd-ambro         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "libft.h"
+
+char	*ft_strcpy(char *dest, const char *src)
+{
+	char	*runner;
+
+	runner = dest;
+	while (*src)
+		*runner++ = *src++;
+	*runner = '\0';
+	return (dest);
+}
